vmware-ose-common-test-suites/requirements.txt
ansi2html==1.5.2
bcrypt==3.1.7
boto3==1.12.46
botocore==1.15.46
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
DateTimeRange==1.0.0
ddt==1.3.1
docutils==0.15.2
Faker==4.0.3
idna==2.9
jmespath==0.9.5
mbstrdecoder==1.0.0
mock==4.0.2
paramiko==2.7.1
pexpect==4.8.0
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
PyJWT==1.7.1
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
shyaml==0.6.1
six==1.14.0
sure==1.4.11
termcolor==1.1.0
text-unidecode==1.3
typepy==1.1.0
urllib3==1.25.9
xmltodict==0.12.0
